pyproject.toml
[tool.poetry]
name = "repl_python3_beta_antispamhysb"
version = "0.1.0"
description = ""
authors = ["Pandapip1"]

[tool.poetry.dependencies]
python = "^3.8"
quart = "^0.15.0"
sklearn = "^0.0"
hypercorn = "^0.11.2"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
